fix(verify): corrige 6 regressões achadas pelo CI (selo de tradução, fuso, i18n removido)
O job `verify` do PR #4 reprovava em 9 asserções, 6 arquivos — todas
consequência de mudanças já commitadas nesta branch, não do merge com
main:

- docs/white-label.en.md: selo de tradução desatualizado (o domínio de
  exemplo mudou em white-label.md). Resselado via
  `pnpm exec tsx scripts/selar-traducao.ts --todas` — a tradução em si já
  estava fiel, só faltava rodar o script.
- lib/leads/timeline-grouping.ts: `Intl.DateTimeFormat("pt-PT", {month:
  "short"})` devolve o mês NUMÉRICO ("20/07"), diferente de pt-BR que
  abrevia por texto ("20 de jul.") — convenção de CLDR diferente entre as
  duas variantes. Troca para `month: "long"` ("20 de julho"), textual nas
  duas, preservando a legibilidade que o próprio comentário do código pede.
- tests/unit/gate-pacing-capability.test.ts,
  tests/unit/pacing-cortesia-vs-antiban.test.ts: fixture `MADRUGADA`
  calibrada para "03h BRT" (UTC-3) parou de cair fora da janela comercial
  7h-22h quando o default virou Africa/Maputo (UTC+2) — 06:00 UTC é 03h em
  São Paulo mas 08h em Maputo, dentro da janela. Recalibrada para 01:00 UTC
  (03h em Maputo).
- tests/unit/handoff-por-orcamento.test.ts: o extrator de rótulo do botão
  procurava `t("...")` — mas o produto removeu a função de i18n (interface
  passou a ser só pt-MZ) antes desta sessão, e o botão virou string literal
  no ternário `{pending ? "..." : "Devolver ao automático"}`. Extrator
  trocado para capturar o padrão atual (achado ao validar, não causado por
  mim, mas bloqueava o CI).
- tests/unit/lgpd-pdf-controlador.test.tsx: mesma causa — o rodapé do PDF
  mudou de "· Relatório LGPD" para "· Relatório de Proteção de Dados — Lei
  n.º 3/2017 (Moçambique)" numa sessão anterior desta branch; o teste ainda
  cortava pelo separador antigo.

As 7 falhas restantes vistas ao rodar localmente eram todas dentro de
`.worktrees/agenda-nativa/` — uma worktree local separada (branch
`fix/install-ia-key-opcional`, já mergeada em main), não relacionada a
este PR e não coberta por CI.

Verificado: as 6 suítes tocadas passam localmente (com env vars de teste
exportadas). `pnpm exec tsx scripts/selar-traducao.ts` confirma "em dia".

+++ lib/leads/timeline-grouping.ts
@@ -103,16 +103,19 @@ export function agrupaTimeline(
   return fino.length <= LIMITE_DE_BLOCOS ? fino : agrupaPorDia(itens, chegouAoVivo);
 }
 
-/** O dia do item no fuso de quem lê — "24 de jul." é o que a pessoa reconhece. */
+/** O dia do item no fuso de quem lê — "24 de julho" é o que a pessoa reconhece. */
 function diaDe(item: TimelineItemView): { chave: string; rotulo: string } {
   const d = new Date(item.performed_at);
   const chave = `${d.getFullYear()}-${d.getMonth()}-${d.getDate()}`;
   let rotulo: string;
   try {
+    // `month: "short"` em pt-PT devolve o mês NUMÉRICO ("20/07"), não abreviado
+    // como em pt-BR ("20 de jul.") — convenção diferente de CLDR entre as duas
+    // variantes. `month: "long"` é textual nas duas, preservando a legibilidade.
     rotulo = new Intl.DateTimeFormat("pt-PT", {
       weekday: "long",
       day: "2-digit",
-      month: "short",
+      month: "long",
     }).format(d);
   } catch {
     rotulo = chave;
